# docs/current/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LightMultiAgentQA",
    "version": "0.7.0"
  },
  "paths": {
...
    "/chat/agno": {
      "post": {
        "tags": [
          "public-default"
        ],
        "summary": "Post Chat Agno",
        "description": "默认基础问答入口：走 agno_chat_service 主链。\n\n路由 async；主链同步函数在线程池执行，避免阻塞事件循环。",
        "operationId": "post_chat_agno_chat_agno_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat/agno/upload": {
      "post": {
        "tags": [
          "public-default"
        ],
        "summary": "Post Chat Agno Upload",
        "description": "拖拽文件入口：multipart 上传文本/常见办公文件，触发文件 prepare 链路。\n\n请求格式：multipart/form-data\n  - message:       用户说的话（如\"先帮我解析这个文件\"）\n  - session_id:    会话 ID（可选，跨轮保存需一致）\n  - use_knowledge: 是否同时检索知识库（默认 false）\n  - file:          支持的文件（≤ 10 MB）\n\n响应：与 POST /chat/agno 完全相同的 ChatResponse 结构。\nextra 字段会额外包含：\n  - pending_kind: \"material_pending\" / \"processing_pending\" / ...\n  - pending_source_id: 当前待保存材料的 pending_id（如存在）\n  - pending_preview: 待保存材料的标题、来源类型与元信息摘要\n\n解析状态说明（extra.pending_preview.metadata.extract_status）：\n  - ok:                 解析成功，资料已进入 pending，可在后续轮次中说「保存到知识库」\n  - unsupported_format: 不支持的文件格式（当前只支持 .txt / .md）\n  - file_not_found:     文件内容为空\n  - parse_failed:       解析过程出错\n  - empty_content:      解析成功但内容为空",
        "operationId": "post_chat_agno_upload_chat_agno_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_post_chat_agno_upload_chat_agno_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}": {
      "get": {
        "tags": [
          "public-default"
        ],
        "summary": "Get Task",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Body_post_chat_agno_upload_chat_agno_upload_post": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Message",
            "description": "用户问题（与 /agno 的 message 字段含义一致）"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话 ID，追问需一致"
          },
          "use_knowledge": {
            "type": "boolean",
            "title": "Use Knowledge",
            "description": "是否启用本地样例知识检索",
            "default": false
          },
          "confirm_long_web_video_asr": {
            "type": "boolean",
            "title": "Confirm Long Web Video Asr",
            "description": "用户已确认对超长网页视频走 ASR（与 JSON /chat/agno 同名字段）",
            "default": false
          },
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "待 prepare 的文件（如 .txt/.md/.docx/.pdf/.xlsx，≤ 10 MB）"
          }
        },
        "type": "object",
        "required": [
          "message",
          "file"
        ],
        "title": "Body_post_chat_agno_upload_chat_agno_upload_post"
      },
...
      "ChatRequest": {
        "properties": {
          "message": {
            "type": "string",
            "maxLength": 32000,
            "minLength": 1,
            "title": "Message",
            "description": "用户问题"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id",
            "description": "会话 ID，追问需一致"
          },
          "use_knowledge": {
            "type": "boolean",
            "title": "Use Knowledge",
            "description": "是否启用本地样例知识检索（V2 最小闭环；仍走 /chat/agno，无新对外主入口）",
            "default": false
          },
          "confirm_long_web_video_asr": {
            "type": "boolean",
            "title": "Confirm Long Web Video Asr",
            "description": "用户已确认对超长网页视频（超过 V16_WEB_VIDEO_ASR_FALLBACK_MAX_SEC）走 ASR",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ChatRequest"
      },
      "ChatResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "task_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Id"
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "answer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer"
          },
          "answer_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Type"
          },
          "task_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Task Status"
          },
          "has_insufficient_info_notice": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Has Insufficient Info Notice"
          },
          "router_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Router Source"
          },
          "primary_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Primary Path"
          },
          "evidence_state": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Evidence State"
          },
          "extra": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra"
          },
          "pipeline_ok": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Ok"
          },
          "debug_stage": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Debug Stage"
          },
          "error_layer": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Layer"
          },
          "pipeline_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Error Code"
          },
          "pipeline_hint_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pipeline Hint Zh"
          },
          "workflow_elapsed_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workflow Elapsed Ms",
            "description": "本轮 chat turn 端到端耗时（毫秒），用于区分慢与错"
          },
          "interaction_mode_zh": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interaction Mode Zh",
            "description": "用户可读的本轮交互模式（如知识库检索、读示例文件等）"
          }
        },
        "type": "object",
        "title": "ChatResponse"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TaskStatusResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok",
            "default": true
          },
          "task_id": {
            "type": "string",
            "title": "Task Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "raw_status": {
            "type": "string",
            "title": "Raw Status",
            "default": ""
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": ""
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "default": ""
          },
          "stage": {
            "type": "string",
            "title": "Stage",
            "default": ""
          },
          "progress": {
            "type": "number",
            "title": "Progress",
            "default": 0.0
          },
          "session_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session Id"
          },
          "request_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Request Id"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "duration_ms": {
            "type": "number",
            "title": "Duration Ms",
            "default": 0.0
          },
          "error_code": {
            "type": "string",
            "title": "Error Code",
            "default": ""
          },
          "failure_reason": {
            "type": "string",
            "title": "Failure Reason",
            "default": ""
          },
          "next_action_hint": {
            "type": "string",
            "title": "Next Action Hint",
            "default": ""
          },
          "result_pending_id": {
            "type": "string",
            "title": "Result Pending Id",
            "default": ""
          },
          "result_source_id": {
            "type": "string",
            "title": "Result Source Id",
            "default": ""
          },
          "result_ttl_seconds": {
            "type": "integer",
            "title": "Result Ttl Seconds",
            "default": 0
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "result_ready": {
            "type": "boolean",
            "title": "Result Ready",
            "default": false
          },
          "pending_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pending Kind"
          },
          "payload_version": {
            "type": "integer",
            "title": "Payload Version",
            "default": 1
          },
          "queue_backend": {
            "type": "string",
            "title": "Queue Backend",
            "default": ""
          },
          "retry_count": {
            "type": "integer",
            "title": "Retry Count",
            "default": 0
          },
          "task_enqueue_to_finish_ms": {
            "type": "integer",
            "title": "Task Enqueue To Finish Ms",
            "default": 0
          },
          "result_status": {
            "type": "string",
            "title": "Result Status",
            "default": ""
          },
          "diagnostics": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Diagnostics"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "status"
        ],
        "title": "TaskStatusResponse"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}